Moved statusCodes.js to util folder for better structure
